object_query controller concern
