+++ app/controllers/admin/scripts_controller.rb
@@ -6,6 +6,7 @@ module Admin
     before_action :set_ga_track, only: %i[show edit update destroy]
     before_action :show_history, only: %i[index]
     before_action :authorization, except: %i[reload]
+    include ObjectQuery
 
     # GET /scripts
     def index
