+++ app/views/admin/layouts/application.html.haml
@@ -4,7 +4,7 @@
   %body.sidebar-mini.skin-blue{ style: 'height: auto; min-height: 100%;' }
     #log
     - if loggedin?
-      .wrapper{ style: 'height: auto; min-height: 100%;' }
+      .wrapper{ style: 'height: auto; min-height: 100vh;' }
         = render 'admin/layouts/navigation', q: @q, appearance: @appearance
         = render 'admin/layouts/sidebar', render: @render
         = render 'admin/layouts/messages'
