+++ app/views/admin/scripts/_listing.html.haml
@@ -4,7 +4,7 @@
       - if can?(Script).destroy_multiple?
         .checkbox
           %label{ for: "checkbox-#{script.id}" }
-            %input{ type: 'checkbox', id: "checkbox-#{script.id}", 'checklist-model' => 'check', 'checklist-value' => script.id.to_s, '@click' => 'listDelete(script.id.to_s' }
+            %input{ type: 'checkbox', id: "checkbox-#{script.id}", 'checklist-model' => 'check', 'checklist-value' => script.id.to_s, '@click' => 'listDelete(script.id.to_s)' }
             %span.cr
               %i.cr-icon.glyphicon.glyphicon-ok
     %td.hide-on-small-and-down
