+++ app/views/admin/users/show.js.haml
@@ -1,4 +1,4 @@
-$("#index-show").html("#{j render('index_show', user: @user)}")
-$("#index-show").css('right', 0)
+$(".index-show").html("#{j render('index_show', user: @user)}")
+$(".index-show").css('right', 0)
 $('.listing table').find(".active").removeClass("active")
 $("#row-#{@user.id}").addClass("active")
